Add reservation start and end time validations
