+++ frontend/src/citizen-frontend/calendar/TimeRangeInput.tsx
@@ -3,35 +3,15 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import React, { useMemo } from 'react'
-import { ErrorKey, regexp, TIME_REGEXP } from 'lib-common/form-validation'
+import { ErrorKey } from 'lib-common/form-validation'
 import TimeInput from 'lib-components/atoms/form/TimeInput'
 import { errorToInputInfo } from '../input-info-helper'
 import { useTranslation } from '../localization'
 
-export interface TimeRange {
+export interface TimeRangeWithErrors {
   startTime: string
   endTime: string
   errors: { startTime: ErrorKey | undefined; endTime: ErrorKey | undefined }
-}
-
-export function validateTimeRange({
-  startTime,
-  endTime
-}: TimeRange): TimeRange {
-  return {
-    startTime,
-    endTime,
-    errors: {
-      startTime:
-        startTime === '' && endTime !== ''
-          ? 'required'
-          : regexp(startTime, TIME_REGEXP, 'timeFormat'),
-      endTime:
-        startTime === '' && endTime !== ''
-          ? 'required'
-          : regexp(endTime, TIME_REGEXP, 'timeFormat')
-    }
-  }
 }
 
 export default React.memo(function TimeRangeInput({
@@ -39,7 +19,7 @@ export default React.memo(function TimeRangeInput({
   onChange,
   'data-qa': dataQa
 }: {
-  value: TimeRange
+  value: TimeRangeWithErrors
   onChange: (field: 'startTime' | 'endTime') => (value: string) => void
   'data-qa'?: string
 }) {
